refactor: extract hook into file
